refactor: use newly renamed api/index
